Move split_image_name() from runner.docker into utils
In preparation for re-using it in the AWS Batch runner.
